+++ apps/raise/src/item/item.service.ts
@@ -440,33 +440,61 @@ export class ItemService {
             /* if current campaign has old image, and the upload process has been done */
             if (i === 0 && imageUpload) {
               if (updateItemData.coverImage) {
-                console.info('Deleting old campaign cover image...');
-                await this.bunnyService.deleteImage(updateItemData.coverImage);
+                console.info(
+                  'Old cover image seems to be exist in the old record',
+                );
+                const isExist = await this.bunnyService.checkIfImageExists(
+                  updateItemData.coverImage,
+                );
+                if (isExist) {
+                  await this.bunnyService.deleteImage(
+                    updateItemData.coverImage,
+                  );
+                }
               }
+              console.info('Cover image has been replaced');
               updateItemData.coverImage = path;
             }
 
             if (i === 1 && imageUpload) {
               if (updateItemData.image1) {
-                console.info('Deleting old campaign image1...');
-                await this.bunnyService.deleteImage(updateItemData.image1);
+                console.info('Old image 1 seems to be exist in the old record');
+                const isExist = await this.bunnyService.checkIfImageExists(
+                  updateItemData.image1,
+                );
+                if (isExist) {
+                  await this.bunnyService.deleteImage(updateItemData.image1);
+                }
               }
+              console.info('Image 1 has been replaced');
               updateItemData.image1 = path;
             }
 
             if (i === 2 && imageUpload) {
               if (updateItemData.image2) {
-                console.info('Deleting old campaign image2...');
-                await this.bunnyService.deleteImage(updateItemData.image2);
+                console.info('Old image 2 seems to be exist in the old record');
+                const isExist = await this.bunnyService.checkIfImageExists(
+                  updateItemData.image2,
+                );
+                if (isExist) {
+                  await this.bunnyService.deleteImage(updateItemData.image2);
+                }
               }
+              console.info('Image 2 has been replaced');
               updateItemData.image2 = path;
             }
 
             if (i === 3 && imageUpload) {
               if (updateItemData.image3) {
-                console.info('Deleting old campaign image2...');
-                await this.bunnyService.deleteImage(updateItemData.image3);
+                console.info('Old image 3 seems to be exist in the old record');
+                const isExist = await this.bunnyService.checkIfImageExists(
+                  updateItemData.image3,
+                );
+                if (isExist) {
+                  await this.bunnyService.deleteImage(updateItemData.image3);
+                }
               }
+              console.info('Image 3 has been replaced');
               updateItemData.image3 = path;
             }
           }
